Put top-level methods inside Git::Pr::Release::Util module
